refactor(settings): extract card sub-components and reduce page size (closes #431)
